fix: fixes failing e2e tests, move saved reply helper functions to a new utility file
